fix: more ux fixes to slack-like query builder
